server/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
pytest
httpx
pytest-asyncio
fastapi-cors
langgraph
langchain-google-genai
langchain-core
